requirements.txt
google-adk
google-genai
fastapi
uvicorn
qdrant-client
pydantic
python-dotenv
pandas
httpx
loguru
anyio
requests
jinja2
